kafka : Connect early and reuse Client
With Kafka, for both consumer and producers, introduce a new constructor
function, and establish a client connection to the broker(s)
immediately.  We use this client in the main consume & produce methods,
but also use the same client in the status call, to ensure we're
checking the actual client connection in use.