Add initial proximo support
This adds initial support for https://github.com/uw-labs/proximo as a
backend as both a source and a sink.

In order to better fit with substrates goals and design, we don't use
the default proximo client here.  This also leads to significantly
better throughput in my testing.

Also included is use of proximo via the substrate url based addressing.