Fix nats status check
There were previously two code paths that indicated in some way that
nats was not connected. One was that the connection was nil, and the
other that the connection was not nil, but IsConnected returned false.
The former returned an error, and the latter returned a non-working
status.

This is not useful to the caller who is expecting an error only if we
cannot establish the status.

Rework and simplify the code so that we always return a status object
(and no error) if we can establish that there is no connection.

Note that returning an error here should only happen if we fail to
establish whether or not things are working, which is actually not
something that ever happens for the nats check, given how the nats api
works.