Avoid calling real straw.Open in test
Instead of calling the real straw.Open, use a mock opener and store the
url we were called with.  This enables us to test that the open function
was called with the expected URL and avoids doing the real open.  Right
now it makes little difference because the real open for `file:///` is
trivial, but if we want to (for example) add tests for our S3 URL
construction, we need this.