Configurable max receive message size
We've published messages into kafka that exceed the 64mb limit
here, so unfortunately this is leading to some services crashing.

Only making this available on the source config, as the sink case
doesn't receive messages that are in anyway bound to the size of
user supplied input.

Rather than adding another argument to dialProximo, I've wrapped
the argumengs up in a struct. Hope that makes sense.