Fix slice index bug
Since needAcks is buffered, we can end up getting an ack before we
realise we need it, leading to a slice index out of bounds.  We could
make needAcks unbuffered, but that will lead to more context switching,
so instead don't assume anything about the relative ordering of acks vs
acks needed.