Save memory in freezer source
Once we have seen an ack for a messages, we discard it from the slice of
messages to ack. However, we do this by slicing the slice to remove the
first element.  This doesn't actually free the message from memory at
that point, so to prevent excessive memory use, also set the data in the
message to nil.