Extend Message to include Key
This PR removes the `KeyFunc` found in the `kafka` sink config, and
replaces it with explicit partition config (mirroring sarama).
Additionally, it extends the `Message` interface to include a `Key()`
function allowing the client to specify it when they are sinking the
message. This way we can avoid the cost incurred by having to deserialize
every single message just to find out which key to use, and we can move
the key identification closer to the business logic, rather than the
initialization of a sink.

- Not all protocols have a concept of partitioning/key functions so
  having `Key` at the high level abstraction isn't the prettiest
- Having to explicitly configure the Hash Partitioner when using `Key`
  doesn't feel great.