Add noop source that blocks until the provided context is canceled.
This is useful when we want to be able to easily disable consumption from one source.
We can use the noop reader and the service will behave as if no messages were coming in.

I have also updated the sink to return `ctx.Err()` instead of `nil` to make it match the
rest of the implementations.